requirements.txt
django
djangorestframework
djangorestframework-api-key
